Refactored AccountPopup to separate component from NavIconButtons
